+++ src/Cli/func/Actions/HostActions/StartHostAction.cs
@@ -33,8 +33,8 @@ using static Azure.Functions.Cli.Common.OutputTheme;
 
 namespace Azure.Functions.Cli.Actions.HostActions
 {
-    [Action(Name = "start", Context = Context.Host, HelpText = "Launches the functions runtime host")]
-    [Action(Name = "start", HelpText = "Launches the functions runtime host")]
+    [Action(Name = "start", Context = Context.Host, HelpText = "Launch the Azure Functions host runtime.")]
+    [Action(Name = "start", HelpText = "Launch the Azure Functions host runtime.", HelpOrder = 3)]
     internal class StartHostAction : BaseAction
     {
         private const int DefaultPort = 7071;